// packages/datadog-ci/src/commands/synthetics/interfaces.ts
export interface MainReporter {
  log(log: string): void
  error(error: string): void
  initErrors(errors: string[]): void
  testTrigger(test: Test, testId: string, executionRule: ExecutionRule, testOverrides: UserConfigOverride): void
  testWait(test: Test): void
  testsWait(tests: Test[], baseUrl: string, batchId: string, skippedCount?: number): void
  resultReceived(result: ResultInBatch): void
  resultEnd(result: Result, baseUrl: string, batchId: string): void
  reportStart(timings: {startTime: number}): void
  runEnd(summary: Summary, baseUrl: string, orgSettings?: SyntheticsOrgSettings): void
}
// ...
export interface Summary {
  /** The ID of the CI batch that was started by this datadog-ci execution. */
  batchId: string
  /** The number of critical errors that occurred during the CI batch. */
  criticalErrors: number
  /** The number of results expected by datadog-ci, prior to any selective rerun. */
  expected: number
  /** The number of results that failed during the CI batch. */
  failed: number
  /** The number of results that failed during the CI batch without blocking the CI. */
  failedNonBlocking: number
  /** The metadata collected for the CI batch. */
  metadata?: Metadata
  /** The number of results that passed during the CI batch. */
  passed: number
  /** The number of results that already passed in previous CI batches on the same commit. */
  previouslyPassed: number
  /** The number of tests that were skipped when starting the CI batch. */
  skipped: number
  /** The public IDs of tests that could not be accessed due to permissions when starting the CI batch. */
  testsNotAuthorized: Set<string>
  /** The public IDs of tests that could not be found when starting the CI batch. */
  testsNotFound: Set<string>
  /** The number of results that failed due to the CI batch timing out. */
  timedOut: number // XXX: When a batch times out, all the results that were in progress are timed out.
}

// Note: This is exposed in CI integrations as a JSON-encoded string in the `rawResults` output.
export interface BaseResult {
  /** The device used in this test run. */
  device?: Device
  /** Duration of this test run in milliseconds. */
  duration: number
  /** The execution rule that was used for this test run. */
  executionRule: ExecutionRule
  /** The ID of the initial attempt if this test run is a retry. */
  initialResultId?: string
  /** Whether this test run is intermediary and expected to be retried. */
  isNonFinal?: boolean
  /** The location from which this test run was executed. */
  location: string
  /** Whether this test run passed, taking into account `failOnCriticalErrors` and `failOnTimeout`. */
  passed: boolean
  /** Raw information about the test run. May not always be present. */
  result?: ServerResult
  /** The ID of this test run. */
  resultId: string
  /** The number of retries, including this test run. */
  retries: number
  /** The maximum number of retries for this test run. */
  maxRetries: number
  /** Information about the selective rerun that was applied to this test run. */
  selectiveRerun?: SelectiveRerunDecision
  /** The test that was run, including any overrides. */
  test: Test
  /** Whether this test run timed out. */
  timedOut: boolean
  /** The timestamp of this test run. */
  timestamp: number
}

// Inside this type, `.resultId` is a linked result ID from a previous batch.
export type ResultSkippedBySelectiveRerun = Omit<
  BaseResult,
  'duration' | 'location' | 'result' | 'retries' | 'maxRetries' | 'timestamp'
> & {
  executionRule: ExecutionRule.SKIPPED
  selectiveRerun: Extract<SelectiveRerunDecision, {decision: 'skip'}>
}

export type Result = BaseResult | ResultSkippedBySelectiveRerun

type Status = 'passed' | 'failed' | 'in_progress' | 'skipped'
type BatchStatus = 'passed' | 'failed' | 'in_progress'

export interface BaseResultInBatch {
  duration: number
  execution_rule: ExecutionRule
  initial_result_id?: string
  location: string
  result_id: string
  retries: number | null
  max_retries: number | null
  selective_rerun?: SelectiveRerunDecision
  status: Status
  test_public_id: string
  timed_out: boolean | null
}

type SkippedResultInBatch = Omit<BaseResultInBatch, 'duration' | 'location' | 'result_id'> & {
  execution_rule: ExecutionRule.SKIPPED
  status: 'skipped'
}

export type ResultInBatchSkippedBySelectiveRerun = SkippedResultInBatch & {
  selective_rerun: Extract<SelectiveRerunDecision, {decision: 'skip'}>
}

export type ResultInBatch = BaseResultInBatch | ResultInBatchSkippedBySelectiveRerun
// ...
export interface TestStepWithUnsupportedFields {
  public_id?: string
  params: {
    element?: {
      multiLocator?: MultiLocator
      userLocator?: unknown
    }
  }
}

export interface TestRequest {
  dnsServer?: string
  dnsServerPort?: number
  headers: {[key: string]: string}
  host?: string
  method: string
  port?: number
  timeout: number
  url: string
}

export interface LocalTestDefinition {
  config: {
    assertions: Assertion[]
    request?: TestRequest
    steps?: {subtype: string}[] // For multistep API tests
    variables: string[]
  }
  locations: string[]
  name: string
  options: OptionsWithUnsupportedFields
  /** Can be used to link to an existing remote test. */
  public_id?: string
  subtype?: string // This is optional in the browser and api schemas
  steps?: TestStepWithUnsupportedFields[] // From browser schema
  type: 'api' | 'browser' | 'mobile'
}

interface Options {
  ci?: {
    executionRule: ExecutionRule
  }
  device_ids?: string[]
  mobileApplication?: MobileApplication
  retry?: {
    count?: number
  }
}

// TODO SYNTH-17944 Remove unsupported fields

export interface OptionsWithUnsupportedFields extends Options {
  bindings?: null | unknown[]
  min_failure_duration?: number
  min_location_failed?: any
  monitor_name?: string
  monitor_options?: any
  monitor_priority?: number
  tick_every?: number
}

// TODO SYNTH-17944 Remove unsupported fields
interface LocalTestDefinitionWithUnsupportedFields extends LocalTestDefinition {
  created_at?: any
  created_by?: any
  creator?: any
  creation_source?: string
  message?: string
  modified_at?: any
  modified_by?: any
  monitor_id?: number
  overall_state?: any
  overall_state_modified?: any
  status?: string
  stepCount?: any
  tags?: string[]
  version?: any
  version_uuid?: any
}
export interface ServerTest extends LocalTestDefinitionWithUnsupportedFields {
  monitor_id: number
  status: 'live' | 'paused'
  public_id: string
}

export type Test = (ServerTest | LocalTestDefinitionWithUnsupportedFields) & {
  suite?: string
}

export interface Assertion {
  actual: string | number | Date | {[key: string]: any}
  errorMessage?: string
  operator: Operator
  property?: string
  target: string | number | Date | {[key: string]: any}
  type: string
  valid: boolean
}
// ...
export interface UserConfigOverride extends BaseConfigOverride {
  /** Override the mobile application version for Synthetic mobile application tests. The version must be uploaded and available within Datadog. */
  mobileApplicationVersion?: string
  /** Override the application version for Synthetic mobile application tests. */
  mobileApplicationVersionFilePath?: string
}
// ...
interface BaseTriggerConfig {
  /** Overrides for this Synthetic test only. This takes precedence over all other overrides. */
  testOverrides?: UserConfigOverride
  /** Name of a test suite (for JUnit reports). */
  suite?: string
}
export interface RemoteTriggerConfig extends BaseTriggerConfig {
  /**
   * Public ID of a test (e.g. `abc-def-ghi`), or its full URL (e.g. `https://app.datadoghq.com/synthetics/details/abc-def-ghi`).
   *
   * @pattern ^(https://.*)?([a-z2-9]{3}-[a-z2-9]{3}-[a-z2-9]{3})$
   */
  id: string
  version?: number
}
export interface LocalTriggerConfig extends BaseTriggerConfig {
  localTestDefinition: LocalTestDefinition
}
export type TriggerConfig = RemoteTriggerConfig | LocalTriggerConfig

export enum ExecutionRule {
  BLOCKING = 'blocking',
  NON_BLOCKING = 'non_blocking',
  SKIPPED = 'skipped',
}

export interface Suite {
  content: TestConfig
  name?: string
}

export interface TestConfig {
  tests: TriggerConfig[]
}

export interface TestSearchResult {
  tests: {
    public_id: string
  }[]
}

export interface APIConfiguration {
  apiKey: string
  appKey: string
  baseIntakeUrl: string
  baseUnstableUrl: string
  baseV1Url: string
  baseV2Url: string
  proxyOpts: ProxyConfiguration
}

export interface APIHelperConfig {
  apiKey: string
  appKey: string
  datadogSite: string
  proxy: ProxyConfiguration
}

export interface DatadogCIConfig extends APIHelperConfig {
  configPath: string
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface SyntheticsCIConfig extends DatadogCIConfig {}
// ...
export interface SyntheticsOrgSettings {
  onDemandConcurrencyCap: number
}

export interface MobileApplicationVersion {
  id?: string
  application_id: string
  file_name: string
  original_file_name: string
  is_latest: boolean
  version_name: string
  created_at?: string
}

export interface ImportTestsCommandConfig extends SyntheticsCIConfig {
  configPath: string
  files: string[]
  publicIds: string[]
  testSearchQuery?: string
}